requirements.txt
asgiref==3.8.1
Django==5.0.2
django-jazzmin==3.0.1
django-taggit==6.1.0
Pillow==10.2.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
mysqlclient==2.2.4
python-dotenv==1.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1
django-cleanup==8.1.0
django-compressor==4.4
django-debug-toolbar==4.3.0
django-storages==1.14.2
requests==2.31.0
six==1.16.0
urllib3==2.2.1
python-magic==0.4.27
django-filter==23.5
django-redis==5.4.0
Markdown==3.5.2